tests/wpt/web-platform-tests/css/css-backgrounds/parsing/border-image-slice-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing border-image-slice with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-image-slice">
<meta name="assert" content="border-image-slice supports the full grammar '[<number> | <percentage>]{1,4} && fill?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("border-image-slice", "1");
test_valid_value("border-image-slice", "1 2% 3 4%");

test_valid_value("border-image-slice", "1 2% 3 4% fill");
test_valid_value("border-image-slice", "fill 1 2% 3 4%", "1 2% 3 4% fill");

</script>
</body>
</html>
